chore!: identified resources can be more cleanly chained
BREAKING CHANGE: An identified resource path will not accept the
identfier in the endpoint function anymore but the api will allow for
resources to be chained. This way it will be closer to the structure of
the REST Api.

For example:
```py
keycloak_admin.roles.by_name("some_role").get()
```